Add timezone support for database datetimes.
This commit adds timezone to all database datetimes, with normalization
to UTC.

Also changes the election mandate period fields in the data models to be
naive `Date`-types.

Note that this breaks the frontend - as the evalg backend now refuses to
handle datetimes without tzinfo. This could be fixed in the frontend by
adding timezone to the submitted datetimes, or we could add some
temporary hacks to accept naive datetimes as local time in the graphene
converters.

Issue: EVALG-419