Fix database fixture parsing
Parsing of json fixtures yields `datetime.datetime` objects for simple
numerical strings.

This commit fixes the parsing, by implementing a custom
`flask_fixtures.loaders.JSONLoader`. This loader is more strict in
accepting `date` and `datetime` strings. To be a `datetime.date`, the
string *must* be on the format `%Y-%m-%d`, and to be a
`datetime.datetime`, the string must be an ISO8601 string with date and
time.