Fix a regression with graphene conversion types
The graphene conversion types must exist when using the flask app.
Refactoring the types caused an issue with using the database from the
cli.

The graphene conversion types are now moved to a separate module, which
is imported by the base graphene module, `evalg.graphql.entities`. That
way, the conversions *should* be loaded whenever any graphene
models are used.