Fix evalg.models module layout
When using SQLAlchemy to build a database schema, all modules that
contains data models needs to be imported - no single import guaranteed
that all models were imported.

This commit re-organizes the layout so that importing evalg.models (or
any sub-module) is enough to make sure that all data models are imported
and available to SQLAlchemy.