Add data models for votes and ballots.
This commit adds an Envelope database model for ballots, along with a
Vote database model for binding ballots to voters.

When adding a Vote object, an additional VoteRecord database model is
created and stored. The VoteRecord model contains fields for logging the
current user and client-ip, with methods for automatically populating
those fields.