Add permission control to fields of ElectionGroup
This is a suggestion for how to permission control the fields of an
ObjectType.
- It is possible to change the default_resolver of a
SQLAlchemyObjectType instance. The default resolver of ElectionGroup has
been changed to a resolver with permission control.
- Custom resolvers need to be decorated with permission_control_decorate
to add permission control for the corresponding field.
- The rules deciding what roles has access to what field, and which target_type
to check, is located in permissions_config.json.
- The newly added test is supposed to fail if all the fields of
ElectionGroup are not permission controlled

Issue: EVALG-590