Add test for verifying permission control
Created a new test which verifies that each of the ObjectTypes specified
in default_config have permission control applied to every one of
its Fields. This ensures that new Fields added to one of those
ObjectTypes will not be available to anyone until a permission
function is added to default_config and the field is decorated with
permission_controller.

Issue: EVALG-590