Except integrity error
This handles an integrity error which would occur if multiple queries
were fired upon a users first login. If the integrity error is
triggered, it means that the user already has a matching person in the
database. If that is the case, the person entry from the database is
used.

Issue: EVALG-620